DAOS-7351 control: misc ras event corrections
